рефакторинг. коэффициенты подобраны ближе к истине

+++ mss_example/SoftConsole/mss_top_sb_MSS_CM3/mss_top_sb_MSS_CM3_app/Modules/MPU6050/mpu6050.h
@@ -389,7 +389,7 @@ void MPU6050_getMotion6(int16_t* ax,
                         int16_t* gx,
                         int16_t* gy,
                         int16_t* gz,
-                        int use_calib);
+                        uint8_t use_calib);
 
 // calibration is not tested!
 void MPU6050_calibration();
